+++ R/utils-time.R
@@ -8,35 +8,45 @@
 #' @param days Number of days of data to include.
 #' @param unit Units used to determine time at end-of-day.
 #'
-#' @description Uses incoming parameters to return a pair of \code{POSIXct}
-#' times in the proper order. The first returned time will be midnight of
-#' the desired starting date. The second returned time will represent the
-#' "end of the day" of the requested or calcualted \code{enddate}.
+#' @description
+#' Uses incoming parameters to return a pair of \code{POSIXct} times in the
+#' proper order. The first returned time will be midnight of the desired
+#' starting date. The second returned time will represent the "end of the day"
+#' of the requested or calcualted \code{enddate}.
 #'
 #' The required \code{timezone} parameter must be one of those found in
-#' \code{base::OlsonNames()}.
+#' \code{\link[base]{OlsonNames}}.
 #'
 #' Dates can be anything that is understood by
 #' \code{lubrdiate::parse_date_time()} including either of the following
 #' recommended formats:
 #'
 #' \itemize{
-#' \item{\code{"YYYYmmdd"}}
-#' \item{\code{"YYYY-mm-dd"}}
+#'   \item{\code{"YYYYmmdd"}}
+#'   \item{\code{"YYYY-mm-dd"}}
 #' }
 #'
-#' @note The second of the returned \code{POSIXct}s will end one \code{unit}
-#' before the specified \code{enddate}. Acceptable units are
-#' \code{"day", "hour", "min", "sec"}.
+#' @section End-of-Day Units:
+#' The second of the returned \code{POSIXct}s will end one \code{unit} before
+#' the specified \code{enddate}. Acceptable units are \code{"day", "hour",
+#' "min", "sec"}.
 #'
 #' The aim is to quickly calculate full-day date ranges for time series whose
 #' values are binned at different units. Thus, if \code{unit = "min"}, the
 #' returned value associated with \code{enddate} will always be at 23:59:00
 #' in the requested time zone.
 #'
-#' @note Because the returned values always satisfy full-day date ranges,
-#' specifying \code{unit = "day"} will cause the returned \code{enddate} to
-#' be associated with 00:00:00 of the following day.
+#' Because the returned values always satisfy full-day date ranges, specifying
+#' \code{unit = "day"} will cause the returned \code{enddate} to be associated
+#' with 00:00:00 of the following day.
+#'
+#' @section POSIXct inputs:
+#' When \code{startdate} or \code{enddate} are already \code{POSIXct} values,
+#' they are converted to the timezone specified by \code{timezone} without
+#' altering the physical instant in time the input represents. This is different
+#' from the behavior of \code{\link[lubridate]{parese_date_time}} (which powers
+#' this function), which will force \code{POSIXct} inputs into a new timezone,
+#' altering the physical moment of time the input represents.
 #'
 #' @return A vector of two \code{POSIXct}s.
 #'
@@ -48,7 +58,6 @@
 #' dateRange("2019-01-08", "2019-01-11", timezone = "UTC")
 #' dateRange(enddate = 20190112, days = 3,
 #'           unit = "day", timezone = "America/Los_Angeles")
-
 dateRange <- function(
   startdate = NULL,
   enddate = NULL,
@@ -57,7 +66,7 @@ dateRange <- function(
   unit = "sec"
 ) {
 
-  # ----- Validate parameters --------------------------------------------------
+  # Validate parameters --------------------------------------------------------
 
   if ( is.null(timezone) )
     stop("Required parameter 'timezone' is missing.")
@@ -65,19 +74,45 @@ dateRange <- function(
   if ( !timezone %in% base::OlsonNames() )
     stop(paste0("Timezone '", timezone, "' is not recognized."))
 
-  # ----- Determine starttime and endtime --------------------------------------
+  if ( !unit %in% c("day", "hour", "min", "sec") )
+    stop("`unit` must be one of: 'day', 'hour', 'min', 'sec'.")
+
+  if ( !is.numeric(days) || length(days) > 1 || days < 1 )
+    stop("`days` must be a single positive number.")
+
+
+  # Handle end-of-day unit -----------------------------------------------------
 
   if ( stringr::str_detect(unit, "^day") ) {
     daySecs <- 60 * 60 * 24
   } else if ( stringr::str_detect(unit, "^hour") ) {
     daySecs <- 60 * 60 * 24 - 3600
-  } else if ( stringr::str_detect(unit, "^min") ){
+  } else if ( stringr::str_detect(unit, "^min") ) {
     daySecs <- 60 * 60 * 24 - 60
   } else if ( stringr::str_detect(unit, "^sec") ) {
     daySecs <- 60 * 60 * 24 - 1
   }
 
-  orders <- c("Ymd","YmdH","YmdHM","YmdHMS")
+
+  # Determine start and end times ----------------------------------------------
+
+
+  # * Prepare POSIXct inputs ---------------------------------------------------
+
+  ## NOTE on hadling POSIXct inputs:
+  #  When given a POSIXct time `lubridate::parse_date_time()` forces the time
+  #  into the timezone given to `lubridate::parse_date_time()`. This alters the
+  #  physical instant in time the original POSIXct represents, so we must
+  #  properly convert a POSIXct start or end date to the proper timezone before
+  #  passing it to `lubridate::parse_date_time()`
+
+  if ( lubridate::is.POSIXct(startdate) ) startdate <- lubridate::with_tz(timezone)
+  if ( lubridate::is.POSIXct(enddate) ) enddate <- lubridate::with_tz(timezone)
+
+
+  # * Parse inputs -------------------------------------------------------------
+
+  orders <- c("Ymd", "YmdH", "YmdHM", "YmdHMS")
 
   if ( !is.null(startdate) && !is.null(enddate) ) {
 
@@ -109,7 +144,7 @@ dateRange <- function(
       starttime <-
         endtime %>%
         lubridate::floor_date(unit = "day") -      # beginning of day
-        lubridate::ddays(days-1)                   # any extra days
+        lubridate::ddays(days - 1)                 # any extra days
     }
 
   } else if ( !is.null(startdate) && is.null(enddate) ) {
@@ -124,13 +159,13 @@ dateRange <- function(
       starttime %>%
       lubridate::floor_date(unit = "day") +
       lubridate::dseconds(daySecs) +             # end of day
-      lubridate::ddays(days-1)                   # any extra days
+      lubridate::ddays(days - 1)                 # any extra days
 
   } else {
 
     # Both missing:  use (now - days), now
     endtime <-
-      lubridate::now(timezone) %>%
+      lubridate::now(tzone = timezone) %>%
       lubridate::floor_date(unit = "day" )
 
     starttime <-
@@ -140,12 +175,14 @@ dateRange <- function(
 
   }
 
-  # Define tlim
+
+  # Order output time limits ---------------------------------------------------
+
   if ( starttime < endtime ) {
-    tlim <- c(starttime,endtime)
+    tlim <- c(starttime, endtime)
   } else {
     # just in case
-    tlim <- c(endtime,starttime)
+    tlim <- c(endtime, starttime)
   }
 
   return(tlim)